Add new fields for zonal isolation (#140)
We want to support workflow zonal isolation in a regional Cadence cluster, which means a workflow started in a zone should be only processed by workers from the same zone even if there are workers of the same task list in other zones.

IsolationGroup field is introduced to poller requests to let pollers only get tasks assigned to the given isolation group.
PartitionConfig field is introduced to workflow trigger requests. Activity and Decision tasks from the workflow are assigned to an isolation group based on this field.