Feature/zonal isolation (#143)
This adds endpoints and structures for the zonal-isolation feature: The ability to subdivide an operating region into more resilient subsections where workflows are confined when they are started.

Internal to Uber and in some companies an 'isolation-group' is sometimes referred to a 'zone', but we are drawing a distinction here, since Cadence can operate sub-zonally with these too.

Workflows are expected to be tagged with the isolation-group they started with and can be partitioned according to whatever partitioning scheme is desired, in the default implementation this will be accomplished by pinning the workflows to the start isolation-group where they were created.

Isolation-group-status:
A enum flag indicating whether an isolation-group is drained or not. For the sake of the initial open-source implementation, only the drained status is important, as anything not specified is expected to be healthy, however, the IDL offers the base for alternative implementations.

Isolation-group-partition:
A single isolation group, with it's name, it's status and potentially any future group-specific information.

isolation-group configuration
An entire set of isolation-groups. This is presented externally as an unsorted list of isolation-group partitions.

APIs for performing actions on Isolation-groups
Apart from engaging in the aforementioned 'pinning' behaviour mentioned above, this PR introduces the ability to perform 'drain' operations on these isolation-groups

This PR also adds the following CRUD endpoints for operating on zones on both a domain and 'globally' (in the sense that it affects all domains).

The APIs are Get / Set apis with upsert semantics. Ie (in pseudocode):
```
// drain zone-1
client.UpdateGlobalIsolationGroups(ctx, adminv1.UpdateGlobalIsolationGroupsRequest{
  IsolationGroups: []*apiv1.IsolationGroupsConfiguration{
      {Name: "zone-1",  Status: apiv1.ISOLATION_GROUP_STATE_DRAINED},
   }
})

// remove all drains
client.UpdateGlobalIsolationGroups(ctx, adminv1.UpdateGlobalIsolationGroupsRequest{
  IsolationGroups: []*apiv1.IsolationGroupsConfiguration{}
})
```