Any-typed global ratelimiting APIs for thrift
This API is purely for inter-Cadence-service use, so the Protobuf equivalent is in the github.com/uber/cadence repository.

Design-wise: this has gone through several iterations, and nearly all top-level keys I can come up with are either totally insufficient or totally unnecessary for one ratelimit design or another, so I've leaned more and more towards "define nothing, allow passing anything".

Obviously this is not super ergonomic, but as the ratelimiter is intended to be pluggable by outside implementations it cannot *wholly* be defined in internal protocol definitions.
Implementers need to maintain in-process definitions on both ends of the protocol, and check type IDs before decoding.  Generally speaking this is probably best done with either a shared protobuf definition (anywhere) or something schema-free like JSON.