Add a usage-data return type to ratelimiters
For compatibility purposes: this will lead to existing requests failing due to an unknown type,
which will eventually put those keys into the "failing" mode.

Which will also be useful to verify this sequence occurs, and we have no true users of this yet,
so it's a harmless "breakage" that we can use to test it.