Add a usage-data return type to ratelimiters (#181)
IDL changes needed to build https://github.com/uber/cadence/pull/6238

Basically this is just: return the "used RPS" that was already being tracked.

For compatibility purposes: while deploying, this will lead to version-crossing requests failing due to an unknown type, which will eventually put those keys into the "failing" mode.
This is currently _intended_, to serve as a full test of that fallback system, but should not be done in general.

Due to a different test (broken hashring) we have already seen the fallback system kick in correctly, so I'm fairly confident that this will work as intended.  Still, it's probably worth testing, particularly because it's easy.

---

This only changes thrift because this type is serialized _outside_ RPC, just like our database types.  There is no proto equivalent, at best it'd be these-serialized-thrift-bytes sent in a proto wrapper.